"ew, php"
Added web version;
Databases are clearly not there --> How does one include a database in a GitHub project? I've always wondered. Probably easily answered with the help of Google, but eh;
I'm pretty sure that's all of the images, but I might be wrong. I still need to make more of them, heh;
Don't judge my code, this wasn't originally intended as a big project (and still isn't, but now it's on GitHub);